requirements.txt
langchain
chromadb
pymupdf
python-docx
beautifulsoup4
requests
sentence-transformers
openai
gradio
